+++ src/js/videojs.record.js
@@ -16,6 +16,7 @@ import RecordIndicator from './controls/record-indicator';
 import PictureInPictureToggle from './controls/picture-in-picture-toggle';
 
 import Event from './event';
+import defaultKeyHandler from './hot-keys';
 import pluginDefaultOptions from './defaults';
 import formatTime from './utils/format-time';
 import setSrcObject from './utils/browser-shim';
@@ -285,6 +286,21 @@ class Record extends Plugin {
 
         // display max record time
         this.setDuration(this.maxLength);
+
+        // hot keys
+        if (this.player.options_.plugins.record &&
+            this.player.options_.plugins.record.hotKeys &&
+            (this.player.options_.plugins.record.hotKeys !== false)) {
+
+            let handler = this.player.options_.plugins.record.hotKeys;
+            if (handler === true) {
+                handler = defaultKeyHandler;
+            }
+            // enable video.js user action
+            this.player.options_.userActions = {
+                hotkeys: handler
+            };
+        }
 
         // hide play control (if present)
         if (this.player.controlBar.playToggle !== undefined) {
@@ -775,7 +791,7 @@ class Record extends Plugin {
         // register starting point
         this.paused = false;
         this.pauseTime = this.pausedTime = 0;
-        this.startTime = new Date().getTime();
+        this.startTime = performance.now();
 
         // start countdown
         const COUNTDOWN_SPEED = 100; // ms
@@ -868,7 +884,7 @@ class Record extends Plugin {
      */
     pause() {
         if (!this.paused) {
-            this.pauseTime = new Date().getTime();
+            this.pauseTime = performance.now();
             this.paused = true;
 
             this.engine.pause();
@@ -880,7 +896,7 @@ class Record extends Plugin {
      */
     resume() {
         if (this.paused) {
-            this.pausedTime += new Date().getTime() - this.pauseTime;
+            this.pausedTime += performance.now() - this.pauseTime;
 
             this.engine.resume();
             this.paused = false;
@@ -1009,7 +1025,7 @@ class Record extends Plugin {
      */
     onCountDown() {
         if (!this.paused) {
-            let now = new Date().getTime();
+            let now = performance.now();
             let duration = this.maxLength;
             let currentTime = (now - (this.startTime +
                 this.pausedTime)) / 1000; // buddy ignore:line